I am making a new branch in this repo to build initial clang prototype for exfil. The only reason it's in here is becuase it will be surrounded by a make configuration and other neat (but ultimately not intended to be useful) files.
Since the state of clang upstream support has been established in the upstream
as being on clang revision 224244, that's what this main exfil branch should be
using until further notice.

as for the actual commit content here, i found this makefile and i was trying to
make it work but it doesnt want to work. That's ok. Might just be some old cruft
or something. the code in here is trivially not interesting, so i'm just
headscratching right now.